Migrated Input unit test to typescript
